Convert save_form_started.jsx to ts
Converts the file and test. It also removes extra param from tests.

[MAILPOET-3523]
